src/requirements.txt
dnspython
cfnresponse